// src/interfaces/client/clientInterface.ts
export interface Client {
    id: number;
    full_name: string;
    first_name: string;
    last_name: string;
    birth_date: string;
    marital_status_name: string;
    cpf: string;
    rg: string;
    mother_name: string;
    email: string;
    mobile_phone: string;
    reference_phone: string;
    reference_responsible: string;
    benefit_type: string;
    situation_status: string;
    benefit_number: string;
    nit_pis: string;
    profession: string;
    ctps: string;
    ctps_series: string;
    inss_password: string;
    contribution_time: number;
    created_at: string;
    non_billable: boolean;
}